+++ FacebookServer/src/server/remote/IFacebookRemote.java
@@ -7,6 +7,6 @@ public interface IFacebookRemote extends Remote{
 	
 	public boolean loginFacebook(String email, String pass) throws RemoteException;
 	
-	public void registrarUsuarioFacebook(String username, String email, String pass, float peso, int altura) throws RemoteException;
+	public void registrarUsuarioFacebook(String username, String email, String pass, int peso, int altura) throws RemoteException;
 
 }
