+++ GoogleServer/src/server/data/UsuarioGoogle.java
@@ -5,10 +5,10 @@ public class UsuarioGoogle {
 	public String username;
 	public String email;
 	public String pass;
-	public float peso;
+	public int peso;
 	public int altura;
 	
-	public UsuarioGoogle(String username, String email, String pass, float peso, int altura) {
+	public UsuarioGoogle(String username, String email, String pass, int peso, int altura) {
 		super();
 		this.username = username;
 		this.email = email;
@@ -39,12 +39,12 @@ public class UsuarioGoogle {
 	}
 
 
-	public float getPeso() {
+	public int getPeso() {
 		return peso;
 	}
 
 
-	public void setPeso(float peso) {
+	public void setPeso(int peso) {
 		this.peso = peso;
 	}
