+++ GoogleServer/src/server/remote/IGoogleRemote.java
@@ -5,6 +5,6 @@ import java.rmi.RemoteException;
 
 public interface IGoogleRemote extends Remote{
 	public boolean loginGoogle(String email, String contrasenia) throws RemoteException;
-	public void registrarUsuarioGoogle(String username, String email, String contrasenia, float peso, int altura) throws RemoteException;
+	public void registrarUsuarioGoogle(String username, String email, String contrasenia, int peso, int altura) throws RemoteException;
 
 }
